// src/types/driver.tsx
// Driver DTO
export interface Driver {
  driverId: number;
  name: string;
  dob: string | null;
  contactNo: string | null;
  fatherName: string | null;
  homeCity: string | null;
  preferredLanguages: string | null;
  martialStatus: string | null;
  spouseName: string | null;
  marriageAnniversary: string | null;
  kids: string | null;
  favouriteSinger: string | null;
  favouriteFood: string | null;
  interestedNews: string | null;
  address: string | null;
  landmark: string | null;
  locality: string | null;
  tehsil: string | null;
  district: string | null;
  state: string | null;
  pincode: string | null;
  assignedLevel: string | null;
}